+++ upcloud/provider.go
@@ -91,6 +91,7 @@ func Provider() *schema.Provider {
 			"upcloud_loadbalancer":                            loadbalancer.ResourceLoadBalancer(),
 			"upcloud_loadbalancer_resolver":                   loadbalancer.ResourceResolver(),
 			"upcloud_loadbalancer_backend":                    loadbalancer.ResourceBackend(),
+			"upcloud_loadbalancer_backend_tls_config":         loadbalancer.ResourceBackendTLSConfig(),
 			"upcloud_loadbalancer_static_backend_member":      loadbalancer.ResourceStaticBackendMember(),
 			"upcloud_loadbalancer_dynamic_backend_member":     loadbalancer.ResourceDynamicBackendMember(),
 			"upcloud_loadbalancer_frontend":                   loadbalancer.ResourceFrontend(),